misc: unify CategoryViewController
